requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-ratelimit==3.0.1
idna==3.3
pymemcache==3.5.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
